feat(37456): Endpoint que consulta status de um username
O novo endpoint retorna a situação do username passado:
- Existente/Não existente no CoreSSO
- Nome e e-mail mesmo que não existente ainda no SigEscola
- Visões do usuário no SigEscola
- Unidades do usuário no SigEscola
- id valido (cpf no caso de não servidor)
