+++ sme_ptrf_apps/users/api/views/user.py
@@ -16,6 +16,7 @@ from sme_ptrf_apps.users.api.serializers import (
 )
 from sme_ptrf_apps.users.models import Grupo
 from sme_ptrf_apps.users.services import SmeIntegracaoException, SmeIntegracaoService
+from sme_ptrf_apps.users.services.validacao_username_service import validar_username
 
 User = get_user_model()
 
@@ -109,6 +110,7 @@ class UserViewSet(ModelViewSet):
             logging.info("Erro ao buscar grupos do usuário %s", erro)
             return Response(erro, status=status.HTTP_400_BAD_REQUEST)
 
+    #TODO Rever url_path 'usuarios/consultar'. É boa prática em APIs Rest evitar verbos. Poderia ser 'servidores'
     @action(detail=False, methods=['get'], url_path='consultar')
     def consulta_servidor_sgp(self, request):
         username = self.request.query_params.get('username')
@@ -125,3 +127,51 @@ class UserViewSet(ModelViewSet):
             return Response({'detail': 'EOL Timeout'}, status=status.HTTP_400_BAD_REQUEST)
         except ConnectTimeout:
             return Response({'detail': 'EOL Timeout'}, status=status.HTTP_400_BAD_REQUEST)
+
+
+
+    @action(detail=False, methods=['get'], url_path='status')
+    def usuario_status(self, request):
+
+        username = request.query_params.get('username')
+
+        if not username:
+            return Response("Parâmetro username obrigatório.", status=status.HTTP_400_BAD_REQUEST)
+
+        e_servidor = request.query_params.get('servidor', 'True') == 'True'
+
+        try:
+            user_core_sso = SmeIntegracaoService.usuario_core_sso_or_none(username)
+            info_core_sso = {
+                'info_core_sso': user_core_sso,
+                'mensagem': 'Usuário encontrado no CoreSSO.' if user_core_sso else 'Usuário não encontrado no CoreSSO.'
+            }
+        except SmeIntegracaoException as e:
+            info_core_sso = {
+                'info_core_sso': None,
+                'mensagem': 'Erro ao buscar usuário no CoreSSO!'
+            }
+
+        try:
+            user_sig_escola = User.objects.get(username=username)
+            info_sig_escola = {
+                'info_sig_escola': {
+                    'visoes': user_sig_escola.visoes.values_list('nome', flat=True),
+                    'unidades': user_sig_escola.unidades.values_list('codigo_eol', flat=True),
+                },
+                'mensagem': 'Usuário encontrado no Sig.Escola.'
+            }
+        except User.DoesNotExist as e:
+            info_sig_escola = {
+                'info_sig_escola': None,
+                'mensagem': 'Usuário não encontrado no Sig.Escola.'
+            }
+
+
+        result = {
+            'usuario_core_sso': info_core_sso,
+            'usuario_sig_escola': info_sig_escola,
+            'validacao_username': validar_username(username=username, e_servidor=e_servidor)
+        }
+
+        return Response(result, status=status.HTTP_200_OK)
